refactor: extract context breakdown computed properties into useContextBreakdown composable
